[TM-1681] Get unit tests passing.
