+++ app/Http/Controllers/V2/Workdays/GetWorkdaysForEntityController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers\V2\Workdays;
 
 use App\Http\Controllers\Controller;
-use App\Http\Resources\V2\Workdays\WorkdayResource;
+use App\Http\Resources\V2\Demographics\DemographicResource;
 use App\Models\V2\Demographics\Demographic;
 use App\Models\V2\Demographics\DemographicCollections;
 use App\Models\V2\EntityModel;
@@ -35,11 +35,12 @@ class GetWorkdaysForEntityController extends Controller
                 // Allows the resource to return an API response with no demographics, but still containing
                 // the collection and readable collection name.
                 $workday['type'] = Demographic::WORKDAY_TYPE;
+                $workday['demographical_type'] = get_class($entity);
                 $workday['collection'] = $collection;
                 $workdays->push($workday);
             }
         }
 
-        return WorkdayResource::collection($workdays);
+        return DemographicResource::collection($workdays);
     }
 }
