+++ app/Http/Resources/V2/Demographics/DemographicResource.php
@@ -9,10 +9,10 @@ class DemographicResource extends JsonResource
     public function toArray($request)
     {
         return [
-            'type' => $this->type,
-            'subtype' => $this->subtype,
-            'name' => $this->name,
-            'amount' => $this->amount,
+            'uuid' => $this->uuid,
+            'collection' => $this->collection,
+            'readable_collection' => $this->readable_collection,
+            'demographics' => empty($this->entries) ? [] : DemographicEntryResource::collection($this->entries),
         ];
     }
 }
