+++ config/wri/linked-fields.php
@@ -483,7 +483,7 @@ return [
                 'pro-rep-rel-paid-nursery-operations' => [
                     'property' => 'workdaysPaidNurseryOperations',
                     'label' => 'Paid Nursery Operations',
-                    'resource' => 'App\Http\Resources\V2\Workdays\WorkdayResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'workdays',
                     'collection' => 'paid-nursery-operations',
                     'option_list_key' => 'workdays-ethnicity',
@@ -491,7 +491,7 @@ return [
                 'pro-rep-rel-paid-project-management' => [
                     'property' => 'workdaysPaidProjectManagement',
                     'label' => 'Paid Project Management',
-                    'resource' => 'App\Http\Resources\V2\Workdays\WorkdayResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'workdays',
                     'collection' => 'paid-project-management',
                     'option_list_key' => 'workdays-ethnicity',
@@ -499,7 +499,7 @@ return [
                 'pro-rep-rel-paid-other-activities' => [
                     'property' => 'workdaysPaidOtherActivities',
                     'label' => 'Paid Other Activities',
-                    'resource' => 'App\Http\Resources\V2\Workdays\WorkdayResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'workdays',
                     'collection' => 'paid-other-activities',
                     'option_list_key' => 'workdays-ethnicity',
@@ -507,7 +507,7 @@ return [
                 'pro-rep-rel-volunteer-nursery-operations' => [
                     'property' => 'workdaysVolunteerNurseryOperations',
                     'label' => 'Volunteer Nursery Operations',
-                    'resource' => 'App\Http\Resources\V2\Workdays\WorkdayResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'workdays',
                     'collection' => 'volunteer-nursery-operations',
                     'option_list_key' => 'workdays-ethnicity',
@@ -515,7 +515,7 @@ return [
                 'pro-rep-rel-volunteer-project-management' => [
                     'property' => 'workdaysVolunteerProjectManagement',
                     'label' => 'Volunteer Project Management',
-                    'resource' => 'App\Http\Resources\V2\Workdays\WorkdayResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'workdays',
                     'collection' => 'volunteer-project-management',
                     'option_list_key' => 'workdays-ethnicity',
@@ -523,7 +523,7 @@ return [
                 'pro-rep-rel-volunteer-other-activities' => [
                     'property' => 'workdaysVolunteerOtherActivities',
                     'label' => 'Volunteer Other Activities',
-                    'resource' => 'App\Http\Resources\V2\Workdays\WorkdayResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'workdays',
                     'collection' => 'volunteer-other-activities',
                     'option_list_key' => 'workdays-ethnicity',
@@ -531,154 +531,154 @@ return [
                 'pro-rep-direct-workdays' => [
                     'property' => 'workdaysDirect',
                     'label' => 'Direct Workday',
-                    'resource' => 'App\Http\Resources\V2\Workdays\WorkdayResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'workdays',
                     'collection' => 'direct',
                 ],
                 'pro-rep-convergence-workdays' => [
                     'property' => 'workdaysConvergence',
                     'label' => 'Convergence Workday',
-                    'resource' => 'App\Http\Resources\V2\Workdays\WorkdayResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'workdays',
                     'collection' => 'convergence',
                 ],
                 'pro-rep-direct-income-partners' => [
                     'property' => 'restorationPartnersDirectIncome',
                     'label' => 'Direct Income Restoration Partners',
-                    'resource' => 'App\Http\Resources\V2\RestorationPartners\RestorationPartnerResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'restorationPartners',
                     'collection' => 'direct-income'
                 ],
                 'pro-rep-indirect-income-partners' => [
                     'property' => 'restorationPartnersIndirectIncome',
                     'label' => 'Indirect Income Restoration Partners',
-                    'resource' => 'App\Http\Resources\V2\RestorationPartners\RestorationPartnerResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'restorationPartners',
                     'collection' => 'indirect-income'
                 ],
                 'pro-rep-direct-benefits-partners' => [
                     'property' => 'restorationPartnersDirectBenefits',
                     'label' => 'Direct In-kind Benefits Restoration Partners',
-                    'resource' => 'App\Http\Resources\V2\RestorationPartners\RestorationPartnerResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'restorationPartners',
                     'collection' => 'direct-benefits'
                 ],
                 'pro-rep-indirect-benefits-partners' => [
                     'property' => 'restorationPartnersIndirectBenefits',
                     'label' => 'Indirect In-kind Benefits Restoration Partners',
-                    'resource' => 'App\Http\Resources\V2\RestorationPartners\RestorationPartnerResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'restorationPartners',
                     'collection' => 'indirect-benefits'
                 ],
                 'pro-rep-direct-conservation-payments-partners' => [
                     'property' => 'restorationPartnersDirectConservationPayments',
                     'label' => 'Direct Conservation Agreement Payment Restoration Partners',
-                    'resource' => 'App\Http\Resources\V2\RestorationPartners\RestorationPartnerResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'restorationPartners',
                     'collection' => 'direct-conservation-payments'
                 ],
                 'pro-rep-indirect-conservation-payments-partners' => [
                     'property' => 'restorationPartnersIndirectConservationPayments',
                     'label' => 'Indirect Conservation Agreement Payment Restoration Partners',
-                    'resource' => 'App\Http\Resources\V2\RestorationPartners\RestorationPartnerResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'restorationPartners',
                     'collection' => 'indirect-conservation-payments'
                 ],
                 'pro-rep-direct-market-access-partners' => [
                     'property' => 'restorationPartnersDirectMarketAccess',
                     'label' => 'Direct Increased Market Access Restoration Partners',
-                    'resource' => 'App\Http\Resources\V2\RestorationPartners\RestorationPartnerResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'restorationPartners',
                     'collection' => 'direct-market-access'
                 ],
                 'pro-rep-indirect-market-access-partners' => [
                     'property' => 'restorationPartnersIndirectMarketAccess',
                     'label' => 'Indirect Increased Market Access Restoration Partners',
-                    'resource' => 'App\Http\Resources\V2\RestorationPartners\RestorationPartnerResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'restorationPartners',
                     'collection' => 'indirect-market-access'
                 ],
                 'pro-rep-direct-capacity-partners' => [
                     'property' => 'restorationPartnersDirectCapacity',
                     'label' => 'Direct Increased Capacity Restoration Partners',
-                    'resource' => 'App\Http\Resources\V2\RestorationPartners\RestorationPartnerResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'restorationPartners',
                     'collection' => 'direct-capacity'
                 ],
                 'pro-rep-indirect-capacity-partners' => [
                     'property' => 'restorationPartnersIndirectCapacity',
                     'label' => 'Indirect Capacity Restoration Partners',
-                    'resource' => 'App\Http\Resources\V2\RestorationPartners\RestorationPartnerResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'restorationPartners',
                     'collection' => 'indirect-capacity'
                 ],
                 'pro-rep-direct-training-partners' => [
                     'property' => 'restorationPartnersDirectTraining',
                     'label' => 'Direct Training Restoration Partners',
-                    'resource' => 'App\Http\Resources\V2\RestorationPartners\RestorationPartnerResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'restorationPartners',
                     'collection' => 'direct-training'
                 ],
                 'pro-rep-indirect-training-partners' => [
                     'property' => 'restorationPartnersIndirectTraining',
                     'label' => 'Indirect Training Restoration Partners',
-                    'resource' => 'App\Http\Resources\V2\RestorationPartners\RestorationPartnerResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'restorationPartners',
                     'collection' => 'indirect-training'
                 ],
                 'pro-rep-direct-land-title-partners' => [
                     'property' => 'restorationPartnersDirectLandTitle',
                     'label' => 'Direct Newly Secured Land Title Restoration Partners',
-                    'resource' => 'App\Http\Resources\V2\RestorationPartners\RestorationPartnerResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'restorationPartners',
                     'collection' => 'direct-land-title'
                 ],
                 'pro-rep-indirect-land-title-partners' => [
                     'property' => 'restorationPartnersIndirectLandTitle',
                     'label' => 'Indirect Newly Secured Land Title Restoration Partners',
-                    'resource' => 'App\Http\Resources\V2\RestorationPartners\RestorationPartnerResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'restorationPartners',
                     'collection' => 'indirect-land-title'
                 ],
                 'pro-rep-direct-livelihoods-partners' => [
                     'property' => 'restorationPartnersDirectLivelihoods',
                     'label' => 'Direct Traditional Livelihoods or Customer Rights Restoration Partners',
-                    'resource' => 'App\Http\Resources\V2\RestorationPartners\RestorationPartnerResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'restorationPartners',
                     'collection' => 'direct-livelihoods'
                 ],
                 'pro-rep-indirect-livelihoods-partners' => [
                     'property' => 'restorationPartnersIndirectLivelihoods',
                     'label' => 'Indirect Traditional Livelihoods or Customer Rights Restoration Partners',
-                    'resource' => 'App\Http\Resources\V2\RestorationPartners\RestorationPartnerResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'restorationPartners',
                     'collection' => 'indirect-livelihoods'
                 ],
                 'pro-rep-direct-productivity-partners' => [
                     'property' => 'restorationPartnersDirectProductivity',
                     'label' => 'Direct Increased Productivity Restoration Partners',
-                    'resource' => 'App\Http\Resources\V2\RestorationPartners\RestorationPartnerResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'restorationPartners',
                     'collection' => 'direct-productivity'
                 ],
                 'pro-rep-indirect-productivity-partners' => [
                     'property' => 'restorationPartnersIndirectProductivity',
                     'label' => 'Indirect Increased Productivity Restoration Partners',
-                    'resource' => 'App\Http\Resources\V2\RestorationPartners\RestorationPartnerResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'restorationPartners',
                     'collection' => 'indirect-productivity'
                 ],
                 'pro-rep-direct-other-partners' => [
                     'property' => 'restorationPartnersDirectOther',
                     'label' => 'Direct Other Restoration Partners',
-                    'resource' => 'App\Http\Resources\V2\RestorationPartners\RestorationPartnerResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'restorationPartners',
                     'collection' => 'direct-other'
                 ],
                 'pro-rep-indirect-other-partners' => [
                     'property' => 'restorationPartnersIndirectOther',
                     'label' => 'Indirect Other Restoration Partners',
-                    'resource' => 'App\Http\Resources\V2\RestorationPartners\RestorationPartnerResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'restorationPartners',
                     'collection' => 'indirect-other'
                 ]
@@ -850,7 +850,7 @@ return [
                 'site-rep-rel-paid-site-establishment' => [
                     'property' => 'workdaysPaidSiteEstablishment',
                     'label' => 'Paid Site Establishment',
-                    'resource' => 'App\Http\Resources\V2\Workdays\WorkdayResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'workdays',
                     'collection' => 'paid-site-establishment',
                     'option_list_key' => 'workdays-ethnicity',
@@ -858,7 +858,7 @@ return [
                 'site-rep-rel-paid-planting' => [
                     'property' => 'workdaysPaidPlanting',
                     'label' => 'Paid Planting',
-                    'resource' => 'App\Http\Resources\V2\Workdays\WorkdayResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'workdays',
                     'collection' => 'paid-planting',
                     'option_list_key' => 'workdays-ethnicity',
@@ -866,7 +866,7 @@ return [
                 'site-rep-rel-paid-site-maintenance' => [
                     'property' => 'workdaysPaidSiteMaintenance',
                     'label' => 'Paid Site Maintenance',
-                    'resource' => 'App\Http\Resources\V2\Workdays\WorkdayResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'workdays',
                     'collection' => 'paid-site-maintenance',
                     'option_list_key' => 'workdays-ethnicity',
@@ -874,7 +874,7 @@ return [
                 'site-rep-rel-paid-site-monitoring' => [
                     'property' => 'workdaysPaidSiteMonitoring',
                     'label' => 'Paid Site Monitoring',
-                    'resource' => 'App\Http\Resources\V2\Workdays\WorkdayResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'workdays',
                     'collection' => 'paid-site-monitoring',
                     'option_list_key' => 'workdays-ethnicity',
@@ -882,7 +882,7 @@ return [
                 'site-rep-rel-paid-other-activities' => [
                     'property' => 'workdaysPaidOtherActivities',
                     'label' => 'Paid Other Activities',
-                    'resource' => 'App\Http\Resources\V2\Workdays\WorkdayResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'workdays',
                     'collection' => 'paid-other-activities',
                     'option_list_key' => 'workdays-ethnicity',
@@ -890,7 +890,7 @@ return [
                 'site-rep-rel-volunteer-site-establishment' => [
                     'property' => 'workdaysVolunteerSiteEstablishment',
                     'label' => 'Volunteer Site Establishment',
-                    'resource' => 'App\Http\Resources\V2\Workdays\WorkdayResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'workdays',
                     'collection' => 'volunteer-site-establishment',
                     'option_list_key' => 'workdays-ethnicity',
@@ -898,7 +898,7 @@ return [
                 'site-rep-rel-volunteer-planting' => [
                     'property' => 'workdaysVolunteerPlanting',
                     'label' => 'Volunteer Planting',
-                    'resource' => 'App\Http\Resources\V2\Workdays\WorkdayResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'workdays',
                     'collection' => 'volunteer-planting',
                     'option_list_key' => 'workdays-ethnicity',
@@ -906,7 +906,7 @@ return [
                 'site-rep-rel-volunteer-site-maintenance' => [
                     'property' => 'workdaysVolunteerSiteMaintenance',
                     'label' => 'volunteer Site Maintenance',
-                    'resource' => 'App\Http\Resources\V2\Workdays\WorkdayResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'workdays',
                     'collection' => 'volunteer-site-maintenance',
                     'option_list_key' => 'workdays-ethnicity',
@@ -914,7 +914,7 @@ return [
                 'site-rep-rel-volunteer-site-monitoring' => [
                     'property' => 'workdaysVolunteerSiteMonitoring',
                     'label' => 'Volunteer Site Monitoring',
-                    'resource' => 'App\Http\Resources\V2\Workdays\WorkdayResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'workdays',
                     'collection' => 'volunteer-site-monitoring',
                     'option_list_key' => 'workdays-ethnicity',
@@ -922,7 +922,7 @@ return [
                 'site-rep-rel-volunteer-other-activities' => [
                     'property' => 'workdaysVolunteerOtherActivities',
                     'label' => 'Volunteer Other Activities',
-                    'resource' => 'App\Http\Resources\V2\Workdays\WorkdayResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'workdays',
                     'collection' => 'volunteer-other-activities',
                     'option_list_key' => 'workdays-ethnicity',
@@ -930,14 +930,14 @@ return [
                 'site-rep-direct-workdays' => [
                     'property' => 'workdaysDirect',
                     'label' => 'Direct Workday',
-                    'resource' => 'App\Http\Resources\V2\Workdays\WorkdayResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'workdays',
                     'collection' => 'direct',
                 ],
                 'site-rep-convergence-workdays' => [
                     'property' => 'workdaysConvergence',
                     'label' => 'Convergence Workday',
-                    'resource' => 'App\Http\Resources\V2\Workdays\WorkdayResource',
+                    'resource' => 'App\Http\Resources\V2\Demographics\DemographicResource',
                     'input_type' => 'workdays',
                     'collection' => 'convergence',
                 ],
